updated INSTALL.sh

+++ INSTALL.sh
@@ -93,13 +93,19 @@ install_on_ubuntu() {
           chmod u+x "$CPLEX_INSTALLER"
           # run the CPLEX installer in a subshell to allow user interaction
           ( ./"$CPLEX_INSTALLER" )
-          rm "$CPLEX_INSTALLER"
-          mv ./ibm/ILOG/CPLEX_Studio2211 "$CPLEX_ROOT"
-          export CPLEX_HOME="$CPLEX_ROOT/cplex"
-          export PATH="${PATH}:${CPLEX_HOME}/bin/x86-64_linux"
-          export LD_LIBRARY_PATH="${LD_LIBRARY_PATH}:${CPLEX_HOME}/lib/x86-64_linux"
-          sh -c "echo '${CPLEX_HOME}/lib' > /etc/ld.so.conf.d/cplex.conf"
-          ldconfig
+          wait
+          if [ $? -eq 0 ]; then
+            rm "$CPLEX_INSTALLER"
+            mv ./ibm/ILOG/CPLEX_Studio2211 "$CPLEX_ROOT"
+            export CPLEX_HOME="$CPLEX_ROOT/cplex"
+            export PATH="${PATH}:${CPLEX_HOME}/bin/x86-64_linux"
+            export LD_LIBRARY_PATH="${LD_LIBRARY_PATH}:${CPLEX_HOME}/lib/x86-64_linux"
+            sh -c "echo '${CPLEX_HOME}/lib' > /etc/ld.so.conf.d/cplex.conf"
+            ldconfig
+          else
+            echo "CPLEX installation failed."
+            exit 1
+          fi
       else
           echo "Error: unable to find the UUID value in the response. The CPLEX download link could not be constructed."
           exit 1
